processor/requirements.txt
kafka-python    # to consume & produce messages in Kafka
pandas          # optional, for transforming or filtering data
